implement another repo approach, alternative to Upgrow's one
