+++ app/repositories/application_record.rb
