--- app/repositories/review_repository.rb
@@ -1,2 +1,0 @@
-class ReviewRepository < BaseRepository
-end
